requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.0
qdrant-client==1.12.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic-settings==2.6.0
pydantic[email]==2.11.7
httpx==0.27.2
python-multipart==0.0.12
sqlalchemy==2.0.36
passlib[bcrypt]==1.7.4
email-validator==2.3.0
